flash-attn不支持FP32,载入模型时需选择BF16/FP16

+++ engine/model_runner.py
@@ -83,7 +83,7 @@ class ModelRunner:
         load_model(model, self.config.model_path)
 
         # 迁移至GPU 并设置为评估模式
-        model = model.to(self.device)
+        model = model.to(self.device, dtype=torch.bfloat16)
         model.eval()
 
         print(f"[ModelRunner] 模型加载完成，设备：{self.device}")
